+++ linux/arch/x86/lib/misc.c
@@ -1,4 +1,3 @@
-// SPDX-License-Identifier: GPL-2.0
 /*
  * Count the digits of @val including a possible sign.
  *
@@ -6,7 +5,7 @@
  */
 int num_digits(int val)
 {
-	long long m = 10;
+	int m = 10;
 	int d = 1;
 
 	if (val < 0) {
